feat: drop hard-coded restrictions for platform-user to prefer whitelists over
Related changes for better support of whitelist for composite commands are in https://github.com/dashpay/dash/pull/6100
