+++ src/rpc/server.cpp
@@ -7,7 +7,6 @@
 #include <rpc/server.h>
 
 #include <chainparams.h>
-#include <node/context.h>
 #include <rpc/blockchain.h>
 #include <rpc/server_util.h>
 #include <rpc/util.h>
@@ -35,10 +34,7 @@ static RPCTimerInterface* timerInterface = nullptr;
 /* Map of name to timer. */
 static Mutex g_deadline_timers_mutex;
 static std::map<std::string, std::unique_ptr<RPCTimerBase> > deadlineTimers GUARDED_BY(g_deadline_timers_mutex);
-static bool ExecuteCommand(const CRPCCommand& command, const JSONRPCRequest& request, UniValue& result, bool last_handler, const std::multimap<std::string, std::vector<UniValue>>& mapPlatformRestrictions);
-
-// Any commands submitted by this user will have their commands filtered based on the mapPlatformRestrictions
-static const std::string defaultPlatformUser = "platform-user";
+static bool ExecuteCommand(const CRPCCommand& command, const JSONRPCRequest& request, UniValue& result, bool last_handler);
 
 struct RPCCommandExecutionInfo
 {
@@ -146,21 +142,6 @@ std::string CRPCTable::help(const std::string& strCommand, const JSONRPCRequest&
         strRet = strprintf("help: unknown command: %s\n", strCommand);
     strRet = strRet.substr(0,strRet.size()-1);
     return strRet;
-}
-
-void CRPCTable::InitPlatformRestrictions()
-{
-    mapPlatformRestrictions = {
-        {"getassetunlockstatuses", {}},
-        {"getbestblockhash", {}},
-        {"getblockhash", {}},
-        {"getblockcount", {}},
-        {"getbestchainlock", {}},
-        {"quorum sign", {static_cast<uint8_t>(Params().GetConsensus().llmqTypePlatform)}},
-        {"quorum verify", {}},
-        {"submitchainlock", {}},
-        {"verifyislock", {}},
-    };
 }
 
 static RPCHelpMan help()
@@ -504,10 +485,10 @@ static inline JSONRPCRequest transformNamedArguments(const JSONRPCRequest& in, c
     return out;
 }
 
-static bool ExecuteCommands(const std::vector<const CRPCCommand*>& commands, const JSONRPCRequest& request, UniValue& result, const std::multimap<std::string, std::vector<UniValue>>& mapPlatformRestrictions)
+static bool ExecuteCommands(const std::vector<const CRPCCommand*>& commands, const JSONRPCRequest& request, UniValue& result)
 {
     for (const auto& command : commands) {
-        if (ExecuteCommand(*command, request, result, &command == &commands.back(), mapPlatformRestrictions)) {
+        if (ExecuteCommand(*command, request, result, &command == &commands.back())) {
             return true;
         }
     }
@@ -539,78 +520,15 @@ UniValue CRPCTable::execute(const JSONRPCRequest &request) const
     if (it != mapCommands.end()) {
         UniValue result;
         const JSONRPCRequest new_request{subcommand.empty() ? request : request.squashed() };
-        if (ExecuteCommands(it->second, new_request, result, mapPlatformRestrictions)) {
+        if (ExecuteCommands(it->second, new_request, result)) {
             return result;
         }
     }
     throw JSONRPCError(RPC_METHOD_NOT_FOUND, "Method not found");
 }
 
-static bool ExecuteCommand(const CRPCCommand& command, const JSONRPCRequest& request, UniValue& result, bool last_handler, const std::multimap<std::string, std::vector<UniValue>>& mapPlatformRestrictions)
+static bool ExecuteCommand(const CRPCCommand& command, const JSONRPCRequest& request, UniValue& result, bool last_handler)
 {
-    const NodeContext& node = EnsureAnyNodeContext(request.context);
-    // Before executing the RPC Command, filter commands from platform rpc user
-    if (node.mn_activeman && request.authUser == gArgs.GetArg("-platform-user", defaultPlatformUser)) {
-        // replace this with structured binding in c++20
-        std::string command_name = command.name;
-        const auto& it = mapPlatformRestrictions.equal_range(command_name);
-        const auto& allowed_begin = it.first;
-        const auto& allowed_end = it.second;
-        /**
-         * allowed_begin and allowed_end are iterators that represent a range of [method_name, vec_params]
-         * For example, assume allowed = `quorum sign platformLlmqType`, `quorum verify` and `verifyislock`
-         * this range will look like:
-         *
-         * if request.strMethod == "quorum":
-         * [
-         *      "quorum sign", [platformLlmqType],
-         *      "quorum verify", []
-         * ]
-         * if request.strMethod == "verifyislock"
-         * [
-         *      "verifyislock", []
-         * ]
-         */
-
-        // If the requested method is not available in mapPlatformRestrictions
-        if (allowed_begin == allowed_end) {
-            throw JSONRPCError(RPC_PLATFORM_RESTRICTION, strprintf("Method \"%s\" prohibited", request.strMethod));
-        }
-
-        auto isValidRequest = [&request, &allowed_begin, &allowed_end]() {
-            for (auto itRequest = allowed_begin; itRequest != allowed_end; ++itRequest) {
-                // This is an individual group of parameters that is valid
-                // This will look something like `["sign", platformLlmqType]` from above.
-                const auto& vecAllowedParam = itRequest->second;
-                // An empty vector of allowed parameters represents that any parameter is allowed.
-                if (vecAllowedParam.empty()) {
-                    return true;
-                }
-                if (request.params.empty()) {
-                    throw JSONRPCError(RPC_PLATFORM_RESTRICTION, strprintf("Method \"%s\" has parameter restrictions.", request.strMethod));
-                }
-
-                if (request.params.size() < vecAllowedParam.size()) {
-                    continue;
-                }
-
-                if (std::equal(vecAllowedParam.begin(), vecAllowedParam.end(),
-                               request.params.getValues().begin(),
-                               [](const UniValue& left, const UniValue& right) {
-                                   return left.type() == right.type() && left.getValStr() == right.getValStr();
-                               })) {
-                    return true;
-                }
-            }
-            return false;
-        };
-
-        // Try if any of the mapPlatformRestrictions entries matches the current request
-        if (!isValidRequest()) {
-            throw JSONRPCError(RPC_PLATFORM_RESTRICTION, "Request doesn't comply with the parameter restrictions.");
-        }
-    }
-
     try
     {
         RPCCommandExecution execution(request.strMethod);
@@ -644,7 +562,7 @@ UniValue CRPCTable::dumpArgMap(const JSONRPCRequest& args_request) const
         // TODO: implement mapping argument to type for composite commands
         if (cmd.first.find(' ') != std::string::npos) continue;
         UniValue result;
-        if (ExecuteCommands(cmd.second, request, result, mapPlatformRestrictions)) {
+        if (ExecuteCommands(cmd.second, request, result)) {
             for (const auto& values : result.getValues()) {
                 ret.push_back(values);
             }
